Applied new database API
