+++ src/Moryx.Runtime.Kernel/Moryx.Runtime.Kernel.csproj
@@ -12,7 +12,7 @@
   <ItemGroup>
     <PackageReference Include="Castle.Windsor" Version="5.0.1" />
     <PackageReference Include="Common.Logging" Version="3.4.1" />
-    <PackageReference Include="CommandLineParser" Version="2.6.0" />
+    <PackageReference Include="CommandLineParser" Version="2.8.0" />
   </ItemGroup>
 
   <ItemGroup>
