+++ src/Moryx.TestTools.SystemTest/Clients/DatabaseMaintenanceWebClient.cs
@@ -3,9 +3,8 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Moryx.Model;
+using Moryx.Model.Configuration;
 using Moryx.Runtime.Maintenance.Contracts;
-using Moryx.Runtime.Maintenance.Plugins;
 using Moryx.Runtime.Maintenance.Databases;
 
 namespace Moryx.TestTools.SystemTest.Clients
