--- src/Moryx.TestTools.UnitTest/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.4.1" targetFramework="net461" />
-  <package id="Castle.Windsor" version="5.0.1" targetFramework="net461" />
-</packages>
