+++ src/StartProject.Core/Config/log4net.config
@@ -1,0 +1,34 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<log4net>
+  <appender name="RollingFileAppender" type="log4net.Appender.RollingFileAppender">
+    <!-- file value is set from log4net.{configuration}.config /-->
+    <appendToFile value="true" />
+    <rollingStyle value="Size" />
+    <maxSizeRollBackups value="10" />
+    <maximumFileSize value="10000KB" />
+    <staticLogFileName value="true" />
+    <layout type="log4net.Layout.PatternLayout">
+      <conversionPattern value="%date - %-5level [%logger].%message%newline" />
+    </layout>
+    <file value="Logs/Runtime.log"/>
+  </appender>
+
+  <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender, log4net">
+    <layout type="log4net.Layout.PatternLayout, log4net">
+      <conversionPattern value="%date [%thread] %-5level [%-50logger] - %message%newline" />
+    </layout>
+  </appender>
+
+  <appender name="TraceAppender" type="log4net.Appender.FileAppender">
+    <file value="trace_log.txt" />
+    <appendToFile value="true" />
+    <layout type="log4net.Layout.PatternLayout">
+      <conversionPattern value="%date [%thread] %-5level %logger [%property{NDC}] - %message%newline" />
+    </layout>
+  </appender>
+
+  <root>
+    <level value="ALL" />
+    <appender-ref ref="RollingFileAppender" />
+  </root>
+</log4net>
