+++ src/StartProject.Core/StartProject.Core.csproj
@@ -14,6 +14,9 @@
     <None Update="app.config">
       <SubType>Designer</SubType>
     </None>
+    <None Update="Config\log4net.config">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
   <ItemGroup>
