+++ src/StartProject/packages.config
@@ -4,6 +4,6 @@
   <package id="Common.Logging.Core" version="3.4.1" targetFramework="net461" />
   <package id="Common.Logging.Log4Net208" version="3.4.1" targetFramework="net461" />
   <package id="log4net" version="2.0.8" targetFramework="net461" />
-  <package id="Microsoft.CSharp" version="4.6.0" targetFramework="net461" />
-  <package id="System.ComponentModel.Annotations" version="4.6.0" targetFramework="net461" />
+  <package id="Microsoft.CSharp" version="4.7.0" targetFramework="net461" />
+  <package id="System.ComponentModel.Annotations" version="4.7.0" targetFramework="net461" />
 </packages>
