+++ src/Tests/Moryx.Model.Tests/packages.config
@@ -5,6 +5,6 @@
   <package id="EntityFramework" version="6.4.4" targetFramework="net461" />
   <package id="NMemory" version="3.1.1" targetFramework="net461" />
   <package id="NUnit" version="3.12.0" targetFramework="net461" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net461" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net461" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net461" />
 </packages>
